python/gRPC/requirements.txt
google-api-core==2.21.0
google-api-python-client==2.149.0
google-auth==2.35.0
google-auth-httplib2==0.2.0
googleapis-common-protos==1.65.0
grpcio==1.67.0
protobuf==5.27.2